Refactor DAT Class
